Add navbar, add auth-base module. Create function for activate module in __init__ file of each module
