+++ manage.py
@@ -1,6 +1,6 @@
 import os
 from wedapp import db, migrate, create_app
-from wedapp.posts.model import User, Post, Comment, Tag, tags
+from wedapp.posts.models import User, Post, Comment, Tag, tags
 
 # Get config Class for app and create app
 env = os.environ.get("WEBAPP_ENV", "dev")
