+++ wedapp/auth/routes.py
@@ -1,0 +1,8 @@
+from flask import Blueprint
+
+auth_blueprint = Blueprint(
+    "auth",
+    __name__,
+    template_folder="../templates/auth",
+    url_prefix="/auth"
+)
