+++ wedapp/main/__init__.py
@@ -1,0 +1,3 @@
+def create_module(app, **kwargs):
+    from .routes import main_blueprint
+    app.register_blueprint(main_blueprint)
