+++ wedapp/posts/__init__.py
@@ -1,0 +1,3 @@
+def create_module(app, **kwargs):
+    from .routes import posts_blueprint
+    app.register_blueprint(posts_blueprint)
