+++ wedapp/templates/navbar.html
@@ -1,0 +1,10 @@
+<nav class="navbar navbar-expand-lg navbar-light bg-light">
+    <a class="navbar-brand" href="{{ url_for('main.home', page=1) }}">Web Blog</a>
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+        <ul class="navbar-nav mr-auto">
+              <li class="nav-item active">
+                <a class="nav-link" href="{{ url_for('main.home', page=1) }}">Home <span class="sr-only">(current)</span></a>
+              </li>
+        </ul>
+    </div>
+</nav>
